requirements.txt
Django==4.2.5  
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.0.0
daphne==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.2
pytest-asyncio==0.21.1
pytest
pytest-django
django-cors-headers
asgiref